chore: moved wrongly ctx_agnostic classified tests in the right place
